refactor: rename peakPicking for getPeaks
