+++ src/getPeaks.js
